added composite rendering
Added RENDERTYPE enum in Configuration, which reads a new key in config
files "renderType". It can declare whether a config should be a volume,
surface, or composite render. If it doesn't exist or is invalid, it
defaults to volume rendering.

The RENDERTYPE can be used in a switch case just like CONFSTATE to
render the correct type. simpleVolumeRender shows an example