added getter/setter for camera width/height
renderer objects now pull the current camera width/height just before
rendering instead of relying on the width and height when the camera was
associated with the renderer. This old behavior is a remnant of when the
renderer could only associate with a single camera and wasn't able to
switch between camera objects. Eventually renderers should just hang
onto an entire pbnj camera object instead of pulling individual pieces
out