light uses camera position instead of view
Previously used view because view was usually set to the negative of the
position. But now with camera view being settable by the config, it's
not necessarily this. However we do want the light emanating from the
camera's position, so we grab it in the renderer when we set the camera
and invert it during the finalization in the render call